Rework GPU event handling
Instead of a list of events, keep a dictionary mapping command queues to
events. This keeps the event list bounded. Importantly, do not clear the
event list in QueueItem.reset, which was a foot-gun.

For some reason this is failing the unit tests. It might be due to a
previously-unseen race condition. The failing test on my laptop is
`test/xbgpu/test_engine.py::TestEngine::test_xengine_end_to_end[NVIDIA GeForce RTX 2060 (CUDA)-4-8192-256-0-heap_accumulation_threshold2]`.