Drastically reduce gc pauses
Use gc.freeze to move the objects allocated at startup to a permanent
generation that is not subject to the GC. This makes GC traversal much
faster, and eliminates some 57ms-long GC pauses.

The fgpu src buffer is reduced to 64MiB (I've tested it as low as 16MiB,
although not for very long).

xbgpu gets the same gc.freeze treatment. I haven't touched the buffer
size since I haven't done any benchmarking on it.