noisy_search: allow more generic probability function
Instead of a fixed probability of an incorrect answer, allow the user to
sample an arbitrary function and pass it. The query point is chosen to
minimise expected entropy rather than simply taking the midpoint (so a
query that provides no information might not be used even if it is in
the middle).