benchmark_fgpu: change calibration test ordering
Iterate over trials as the outer loop and rates as the inner loop. This
will be slower (dsim has to be restarted each time), but will hopefully
give more statistically valid results since any effect that slows down
the machine for a few minutes will be spread over all the rates rather
than concentrated on a single rate.