Write tests for postproc 4-bit output
I also sprinkled on a few type annotations while I was at it, which in
turn revealed that there were bugs in the parameter order to
postproc_host (two bugs which cancelled out).

To simplify the test, the postproc_host function no longer returned an
integer dtype. Instead, it quantises and saturates, but leaves the
values as complex64, and the received data is unpacked to complex64 for
comparison.

Closes NGC-1099