WIP Update BSend for BPipeline
As well as surrounding loop infrastructure in BPipeline. A lot of the
updates are placeholders/indicators for where more targeted work should
be placed.

Currently the BSend only transmits zeroed data, the timestamps are yet
to be figured out and stream configuration to be tweaked. Still, it's a
first pass.

Contributes to: NGC-1011.