benchmark_fgpu: use single-pol dsims when n==2
This is needed for Genoa tests, as a single thread isn't sufficient to
hit maximum rates with n==2.

This also fixes a bug where --first-id was being passed incorrectly to
dsim when it runs two streams. It doesn't matter at all though (all that
matters is that the two pols differ in the last bit of the digitiser
ID).