noisy_search: fix expected entropy calculation
It was using the probability of the query value being more/less than the
true value to determine the probability of getting a yes/no outcome,
which ignores the noise. We already calculate the actual probability of
a yes/no value as part of normalisation, so we just use that.