Some refactoring on memcpy_loop
Write a common function to handle alignment for SSE2/AVX/AVX512 copy
implementations. It is more standards-conforming than the previous
implementation, because it doesn't rely on the object representation of
pointers (uses std::align instead).

Also added self-test functionality to make it practical to verify that
the copy functions actually work as advertised.