Overhaul BPipeline send to support multiple --beam
This includes a contained, but invasive update to the capture-{start,
stop} mechanism in XBEngine. There is also a plumbing change in BSend to
accommodate for comments on PR #650, as well as the multiple BOutput
update.

Naturally, an update to test_bsend was required.

Contributes to: NGC-931, NGC-1011 and NGC-1139.