New approach to figures in qualification report
Instead of generating LaTeX code (with the pgf backend), generate PDFs
and embed them in the JSON file with base64 encoding, then use
`\includegraphics` to insert them into the report. These use a
`$msg_type` of `binary_figure` so that old reports can still be
processed. Old reports now extract the figure code to a temporary file
which is included with `\input`, which should also make things more
robust (because the surrounding environments will be able to find the
end of the command without becoming confused by category code changes in
the generated code).

This will hopefully address some of the performance issues identified in
NGC-1196 and the robustness issues plaguing NGC-1182.