Add second test of beamformer delays
It does not try to cancel out the delay in the F-engine, but simply
checks the results against expected values. It also shows the effect on
phase of the delays (which does reveal some artifacts due to the lack of
dithering in the beamformer quantisation).

Also change the default matplotlib style to use thinner lines, instead
of overriding it on a case-by-case basis.