Improvements to layering in the docker build
The main improvement is that the package itself is managed in a separate
layer to the dependencies, so that changes to the source code will
only make small new layers to push instead of a fresh copy of all the
dependencies. That's achieved by having a build layer generate a wheel,
which is copied into the final FROM and installed. That does waste some
space in that final FROM (the wheel is deleted afterwards but still
requires a layer).

Also simplified the way the right version of pycuda is installed, and
use COPY --link because the docs say that's what one generally wants.