Cache HeapReferenceList in bsend
This avoids having to recompute the list of heaps on every transmission.
Instead, it is recomputed (on each frame) whenever it changes, which is
much rarer.

This makes a noticeable (~20%) improvement to CPU usage when there is a
large number (50+) beams, although at present the beamformer kernel is
the bottleneck to supporting more beams.