Make beamform saturation tests more deterministic
The beamformer dithering made it impractical for the CPU to exactly
determine which values saturate, and the previous approach was to just
use an approximate test (relying on the central limit theorem, but
without any rigorous statistics to determine standard deviation).

Instead, compute lower and upper bounds on the possible number of
saturated values. That's not a particularly strong test, but any
failures are serious.

And there are failures. I suspect NGC-1172 (missing data the end is
still run through the pipeline using old values).