Add a sync-time sensor for dsim
This will make it possible for a user to actually learn the sync time of
a subarray product they've just spun up, even if it doesn't have any
connected antenna-channelised-voltage stream (which reports its sync
time via code in katsdpcontroller).

To make this reliable, the sync time is chosen slightly earlier in the
startup: specifically, before starting the katcp server, so that a
client that immediately queries the sensor will get an answer rather
than ever witnessing the 'unknown' status. It is still kept after
generation of the signal data to minimise the starting timestamp when
the sync time is implicit.