Overhaul xbgpu/test_engine for non-contiguous data
In order to verify the update to the BPipeline for handling missing
data, there were some tweaks in the previous commit to the XPipeline,
both of which culminate in this rework of the end-to-end test.

There was some reshuffling of data verification as mypy complained the
test was too complex.

Contributes to: NGC-1172.