Qualification: change NB centre freq
The change to spectra-per-heap in NGC-1121 exposed a weakness in the
baseline_correlation_products consistency test: for narrowband it only
works if the dump rate is an integer fraction of the centre frequency.
Reducing spectra-per-heap gave more resolution to the dump rate, and
this was no longer the case.

Fix by choosing a coarser centre frequency for narrowband.

Closes (perhaps only partially) NGC-1304.