Extend timeout of qualification tests to 7 hours
See NGC-1318.

I also shifted the scheduled run two hours earlier (i.e. 01:00 SAST rather than the 03:00 that it previously was) to accommodate this.

I still want to get to the bottom of why it's suddenly not finishing inside the timeout, but if an extra two hours doesn't work then it's probably quite egregious.