Make test_group_delay fail fast if data is missing
Since it's using complete_chunks() directly instead of
next_complete_chunk(), it doesn't have any timeout, and if complete
chunks aren't arriving it can hang the test forever.

Also fixes a bug where if non-contiguous chunks are seen, it could hang
things by not returning the chunk to the stream. It's possible that this
is the real reason for NGC-1341.

Closes NGC-1341.