Set dig-rms-dbfs sensors to FAILURE on missing data
If there's missing digitiser data in a given set of output spectra, then
calculating the average digitiser power level isn't particularly
meaningful.

This commit introduces a check for if all the data is present first, and
if it is, then the sensor is updated with the average power as
calculated on the GPU. If there is missing data though, the sensor is
set to a FAILURE state.

Closes NGC-855 (if it works).