Partially unify item generation for fgpu.send
Both data and descriptor heaps now start from a new make_item_group
function. They set the dtype/format independently, because the logical
type that should be used for decoding is not necessarily accessible as a
numpy dtype.

Since typing-extensions is now a direct dependency, I took the
opportunity to update it to the latest version. It works fine with the
previous version, but there are apparently performance and compatibility
enhancements in later versions.

See NGC-1156.