Replace -np.inf with minimum float64 value
CAM's proxies have problems with non-finite sensor values for some
reason that I can't remember (possibly JSON encoding).

I am reluctant to make the value 0 even though the status is failure and
therefore the value is meaningless, because I'm not sure whether
Prometheus / Grafana considers the value when scraping things. 0 is the
top of the dbfs range, while a negative infinity will be the bottom of
it. If we're going to see spikes in the graph for missing packets, I'd
rather that they are spikes downward than spikes upward.