Improve accuracy of qual test timestamps
Timestamps reported during the test are expressed in relative terms by
subtracting a start time. The start time was determined by calling
time.time() in the pdf_reporter fixture setup, but it's possible for
other fixtures to be set up between that point and the start of the test
itself (particularly the cbf fixture, which can take several seconds to
set up).

Fix it by adding a pytest_runtest_call hook which runs right before the
test actually starts (modulo things like setting up log capture, which
are also handled with that hook) and replacing the test_start field with
a more accurate version.

Relates to NGC-1343 (doesn't fix anything, but will make analysis
easier).