Add unit testing for dig-rms-dbfs sensor updates
If there are a handful of missing packets, the test now passes.

However, in the test_missing_packets test there is a large range where
things are just dropped and for some reason the sensor is going for a
very long time without updating. I'm not sure how to solve it from here.