Qualification: rework XBReceiver
WIP - DO NOT MERGE

A worker task now receives complete chunks and decides to either enqueue
them or discard them. This should reduce backpressure on the spead2 receiver
and hence incomplete heaps. However, it still happens because the worker is
an async task rather than a separate thread and so doesn't get to run when the
test is performing computations. Using a thread may be better but it would
have more complex communication to ensure safe access to the queue.

This didn't seem to make any improvement to the reliability issues I was
having with NGC-938, but it may prove useful for NGC-1085.