Fix test_missing_heaps for long gaps
The checking of the dig-rms-dbfs sensor as I was previously doing it was
not handling long gaps of more than an output chunk, which there are in
the unit test. So this update fixes that.

I also exclude narrowband from checking the sensor, because it's only
going to be present in the first wideband output.