Fix a subtle bug in the out_item.present handling
In the case where some heaps were missing precisely at the end of a
chunk, the entire chunk would slip through without realising that the
last one(s) were missing, because in-between transmissions the `present`
array was not actually "reset" in any meaningful way.

This commit fixes that, and with that all the NGC-855 work is done,
pending whatever code cleanup feedback I now get.