Make qualification reporting more robust to failed CBFs
When a CBF fails to configure, the tests don't even get to the point of
reporting their blurbs, which was leading to errors about inconsistent
names/blurbs. We should probably tweak the conftest to fix that, but for
now this makes the reporting more robust so that we can at least get a
PDF out the other end.

Relates to NGC-938.