Qual: move test info setup into a hook function
Setting it up in the pdf_reporter fixture means it didn't get set up if
an earlier/coarser-scoped fixture (typically the cbf fixture) failed.
With a hook function, we can ensure it is populated before we start
setting up any fixtures.

This does mean it will get configured for *every* test, even if the test
doesn't use the pdf_reporter fixture. That is probably the behaviour we
want though, since otherwise tests might be completely omitted from the
report.