Reset fgpu out_item.present between chunks
This hadn't been done previously, so it left the possibility, if the
user wasn't careful to compare n_spectra and capacity, that you could
think there are more heaps present than there actually are.

This resets it.

I'm basically just committing so that Jenkins can build and I can test
the performance impact of this, I suspect there may be more work later.