Do not update rx.synchronised when there is no input data
If an accumulation has no data at all, do not do any processing on it,
including updating rx.synchronised. This gives more consistent
behaviour: previously rx.synchronised could get updated if an input
chunk overlapped the accumulation, even if it didn't contain any valid
heaps for that accumulation. With this change, the observable behaviour
no longer depends on the chunk size.

This allows the testing for rx.synchronised to be made more robust
(checking the exact order of updates).