Make use of typing.Self
This makes type annotations more accurate because it indicates that the
method returns the type of the instance on which it is invoked, rather
than of the class in which it is defined. monitor.py already indicated
this but used an older, more roundabout idiom.

Note that the binary operators (__add__ etc) in Signal do NOT return
Self: contractually they return an unspecified subclass of Signal, and
both contractually and in practice it won't always be the same class as
the left-hand side.