WIP on making use of TaskGroup
Where asyncio.gather was used to start, run and await multiple tasks,
use the new TaskGroup instead, which has better cancellation behaviour
(cancels the other tasks as soon as one task fails, and potentially
report multiple exceptions via an ExceptionGroup).

This is not used everywhere it could be yet. In particular, tasks
started by engines are managed by aiokatcp's add_service_task, which
does similar things to TaskGroup (although it does not raise exception
groups). There are a number of other places where the creation of and
waiting for a task are widely separated such that it would be
inconvenient to use a context manager.

See NGC-1380.