Make test_dig_rms_dbfs_sensors more robust
The window size is set to 2 *output* chunks, but the expected timestamps
in the test were computed based on the *input* chunk size. If they
weren't equivalent, the test would fail.

The calculation of the window size also failed to account for narrowband
decimation. Fixing that required some restructuring of the code, since
the engine_server fixture wasn't always used with a specific Output.
There is now a dedicated function in Pipeline that is intended for
monkeypatching.

See NGC-1369.