Make global memory access for dithering more efficient
Save and restore only the state that's actually needed. This reduces
memory usage and global memory traffic for beamforming. In some
microbenchmarks it makes beamforming 1.6x more efficient. It also
simplifies some of the code by eliminating the need to measure the size
and alignment of curandStateXORWOW_t at runtime.

Also removed a redundant initialisation of random state in
beamform_bench.py.

This is indirectly related to NGC-1264: this improvement is needed to
make F-engine dithering feasible.