Disable dithering - DO NOT MERGE
This is a quick hack that replaces the dither function to return 0, just
for testing the effect. A permanent removal should go deeper (completely
ripping out the infrastructure for it, and restoring tests to their
tighter limits), and if it's going to be optional it should be controlled
by a switch.