Add dither option to xbgpu
The kernel code is more complex than for fgpu, because one kernel
handles multiple beams, and they could have individual choices of
dithering. This is already handled for the polarisations, so we use the
same pattern: the choice is baked into the BeamformTemplate, and encoded
into the kernel source as a constant array generated at compile time.
The indices into the array should also be known to the compiler
(assuming it unrolls loops well) and hence it shouldn't add any cost.
In addition to this mechanism, if none of the beams have dithering, we
remove the random state array entirely.

Closes NGC-1429.