Add ability to disable descriptors on dsim
During the course of NGC-1197 I battled to get SKARAB F-engines to
accept data from the softare D-sim. The current suspicion is that it has
to do with the descriptors which are issued on a regular basis - it's
not a data packet so it screws with the FPGA's syncing logic.

In this commit I add the ability to disable the descriptors to see
whether that helps.