Qual test_control: manually start recv streams
This allows the streams to start up with no backlog of data in them that has
to be caught up with. It also means the host doesn't show RDMA out-of-buffer
errors at the start of the test (there are still some during shutdown).

Currently it's only deployed for test_control, but there are probably other
tests that could benefit from it. It could even become the default, with
complete_chunks() automatically starting the stream if needed.

Relates to NGC-1265.