Add an optional timestamp argument to capture-start
Heaps that preceed the timestamp are not transmitted. This makes the
versioning of send_enabled for BSend invalid; it is replaced by using
the set of heaps to send as a cache key. It's unclear whether this
caching has much benefit at that point, but it does still save a bunch
of transitions through the pybind11 boundary.

There are no unit tests for this yet.

See NGC-1462.