Improve SNR in test_group_delay
The original test had a very large standard deviation for the measured group
delay, which made the test quite weak (and indeed it missed a bug in the
calculation of reported delay). Improve it by putting tones in multiple
channels and measuring them all, to increase the amount of data being
averaged.

There are two issues with using the sum of a large number of tones in the
dsim:
- it's expensive (multicw doesn't currently use an FFT, and it might not be
  practical from a memory usage standpoint).
- the amplitude has to be reduced in proportion to the number of tones to
  prevent overflow.

Distributing the tones across the dsims mitigates both issues. Increase the
number of dsims for antenna-channelised-voltage tests from 24 to 32.

See NGC-1464.