Try testing beams 0, 2, 4 instead of 0, 1, 2
The working hypothesis that this is meant to test that, since the
F-engines dither H and V pols differently, this might end up resulting
in the failure we were seeing for NGC-1473.

If that is indeed the case, we should be able to solve it by comparing
like-to-like pols, and this commit will test that.