Compare like pols in test_gain
The qualification/tied_array_channelised_voltage/test_gain test
occasionally shows a difference one greater expected for a single
value in the test every now and then. It's very intermittent. To a first
approximation this shouldn't be possible because the tolerance for each
test should be taken into account and there's no way for dithering to
cause a bigger difference assuming that the incoming signal is the same.

We think that this is caused by separate dithering for each pol in the
F-engine which means that despite having the same d-sim input, the
F-engine output for pol0 and pol1 may be ever so slightly different from
each other, just different to cause the error.

In this commit I adjust the test to using the same pol in three
different beams for comparison, rather than three half-beams which would
have been pol0, pol1, pol0.

This appears to solve the problem.

Closes NGC-1473.