Change ownership of /venv in Jenkins docker image
Jenkins, when installing stuff via pip in the container, needs to have
the venv writeable. After the container is built, the directory is owned
by root. This changes it to be owned by the UID:GID 1000:1000 (which has
a different name every time, in the container it appears to be `ubuntu`
but that shouldn't be as important.