Modify qualification Jenkinsfile for saner agents
In the Jenkins setup that our former colleague put together, we had
"clouds" rather than Nodes, which ran their containers as root. It was a
rather complicated setup and we've simplified things, but in the new
regime containers don't get run as root anymore so they need slightly
different capabilities.

This commit adds spead2_net_raw into the Jenkins image so that the agent
that runs qualification can use its magic to enable ibverbs. In order
for that to work, we need to install a couple more libraries in the
`jenkins` docker image. We also add usage of the spead2_net_raw
executable to endow pytest with ibverbs.

Contribures to NGC-1447.