Reduce noise in test_group_delay
The test uses the beamformer to combine results from multiple antennas. What's
actually needed is a mux, but beamformer does a sum. In theory there should
only be signal in one antenna for each channel, but spectral leakage seems to
be adding noise, and it doesn't seem to be entirely random (it causes bias in
the distribution of the final phases).

Use F-engine eq gains to suppress this noise and ensure we really are just
muxing together the data we want.

Hopefully closes NGC-1554.