Switch around a confusing if statement
The statement involved a double negative. (Un)negate the test and
switch the if and else arms.

This ran into a mypy issue: it infers the type of tone_channel from the
'if' arm, even if it's annotated as type: ignore there. Fix by writing
the expression in an equivalent way that doesn't mess up the type
annotation.