Change stop-band calculation for _generate_ddc_weights_no_discard
Instead of trying to get zero response starting from the Nyquist
frequency, allow aliasing into the "unused" part of the band.

By relaxing the constraints, we need fewer taps to get decent filter
response. So reduce the default number of taps as well. Note that this
will need a matching change in katsdpcontroller.

See NGC-1399.