Specify strict= parameter to zip() calls
Add target-version="py312" parameter to ruff configuration. This enables
it to give warnings that are not appropriate to older Python versions.
The primary is that it recommends specifying the 'strict' parameter to
zip() calls to indicate whether the inputs are expected to be the same
length. This is probably good documentation and a good sanity check so
I've done that.