Speed up sending of zero signals
Keep a third HeapSet which is initialised to all zeros, and switch to
using it when an all-zero signal is requested. This should have less
overhead than my previous attempt, which was done at the signal.py layer
and had a fast path for generating an all-zero signal (but still
required actually writing a bunch of zeros to memory).