Qualification: reuse receivers
This reduces the startup time for each test, because it doesn't need to
recreate all the receiver infrastructure each time.

Tests that used the manual_start versions of the receivers now just use
the regular version; the manual_start variant was essentially just an
optimisation that is no longer needed.

Relates to NGC-1704.