Qualification: cache receivers between tests
This reduces the startup time for each test, because it doesn't need to
recreate all the receiver infrastructure each time. However, it will
probably break tests that explicitly use the _manual_start variants of
the receivers, because now it'll just depend on whether the receiver
had already been started by the previous test.