Eliminate XBReceiverCache
Instead of starting streams on demand, always run both BCP and TACV receivers
for the lifetime of the subarray product.

This will make it easier in future to collect and report statistics about
lost data over the life of a subarray product.

Additionally, eliminate the manual_start variants of the fixtures, which are
not compatible with always-on receivers.