Qualification: batch additions to ChunkQueue
This seems to make it much more reliable, at least on test_control. My guess is
that when the worker thread has been locked out for some time by the GIL, this
allows it to quickly throw out most of the backlog, instead of waking up the
main thread on the first (out-dated) heap.

There are still occasional failures for the TACV data. There don't seem to be any RDMA
overruns on the receiver machine, so perhaps the ChunkQueue is too small.