Fix stats collection for stream groups
The way StatsCollector uses weak references was incompatible with
ChunkStreamGroupMember: the latter is a thin Python wrapper that is
created on the fly to wrap the real C++ class, and which immediately
vanishes if the Python code doesn't hold a strong reference.

Fix it by using a proxy class that behaves like a weakref to a
ChunkStreamGroupMember, but actually holds a weakref to the
ChunkStreamRingGroup plus an index.

There is a unit test added as a basic sanity check, although it doesn't
actually test this subtlety because can't easily be done with mocking.
I've done a manual test (with a deliberately-broken dsim) to check that
the fix is working.

Closes NGC-1468.