Add add_send_arguments helper function
While implementing this, I noticed that dsim didn't have a --comp-vector
argument, and is defaulting to using comp vector 0 for all dsims. I
added the argument, but it should get wired into katsdpcontroller.

xbgpu had a different default for --send-comp-vector (1 instead of 0).
Since we always specify the comp vector when doing anything serious, I
decided not to try to preserve that.