Update dependencies to latest versions
With the exception of the CUDA base Docker image: I upgraded to 12.9.1
rather than 13.x because that would require a driver upgrade on the
hosts.

This required a few code changes because numpy now has more precise
typing, which was causing mypy to know more about what type something is
and complain, even in cases where we don't actually care about the
different between Python floats and numpy floats.