Rewrite parsing of --wideband and friends
--wideband, --narrowband, --corrprod and --beam all use key=value pairs.
Previously, the parsing code was not shared between xbgpu and fgpu, and
wasn't the most maintainable (with information about the type, the
default and whether it was required spread around). Replace it with
unified code in katgpucbf.main that presents a simplified version of the
ArgumentParser interface.

The new code also raises ArgumentTypeError on errors rather than
ValueError, so that the error messages are more helpful, instead of just
saying that the value is invalid.

Testing still needs to be improved: a lot of the existing tests can be
deleted and replaced by tests on the common code, which should reduce
duplication.