Unify iter_chunks between engines
Recycle the code from fgpu and generalise it to handle xbgpu as well.
This has a couple of user-visible side effects:

- xbgpu gains a input_samples Prometheus counter
- xbgpu no longer logs a warning when a chunk is entirely skipped. We
  get sufficient information from the Prometheus counters, and tend not
  to read those logs.

This also includes some fixes and changes to the Layout classes, which
were confusing and inconsistent: specifically, in fgpu, chunk_heaps was
actually the number of heaps per polarisation per chunk. It's been
renamed chunk_batches and now chunk_heaps is actually what it says on
the tin. The glossary is also updated to define "batches".

Rename chunk_heaps_time -> chunk_batches