benchmark_*gpu: be more robust to high latency connections
To measure the amount of data received, we
- query Prometheus over HTTP (from the machine running benchmark_*gpu)
- sleep for N seconds, in parallel with the above query
- query Prometheus again

That process ensures that we start the queries exactly N seconds apart,
even if the query takes non-trivial time. However, the first query had
to establish the HTTP connections (TCP handshake), while the second
reuses the existing connections; that means that the first would take
longer, and we'd still end up measuring less than N seconds of data.

This is still vulnerable to jitter, but if there is consistent latency
it should be a lot more accurate.